moved project implementation into src/project, src/dtl/project was kinda meaningless
